fix: use multiaddr as class name
Cosmetic change ahead.

When logging multaddrs in browsers `DefaultMultiaddr` appears in
the console.

It's called this to prevent a collision with the `Multiaddr`
TypeScript interface name.

The interfaces are removed at compile time so here we import the
interface with a different name, freeing up the `Multiaddr` symbol
to be used as the class name.
